// src/types/post.ts
// src/types/index.ts
export interface Author {
  name: string;
  avatar: string;
  role: string;
}

export interface Post {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  category: string;
  author: Author;
  image: string;
  date: string;
  readTime: string;
  featured: boolean;
  tags: string[];
}